server/requirements.txt
Django>2.2,<2.3
click
django-cors-headers==2.2.0
django-silk==3.0.2
djangorestframework-camel-case==1.0.3
djangorestframework==3.8.2
drf-dynamic-fields==0.2.0
psycopg2
pyjwt==1.5.3
requests==2.13.0
shapely==1.6.4
